# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Water Pump Monitoring System (SWPMS) API",
    "description": "\n### PT Ascon Multi Pratama - Industrial IoT & Water Pump Monitoring Dashboard\n\nSWPMS is a centralized, real-time, and historical water pump monitoring backend platform.\nThis API provides full operational telemetry queries, dynamic site/area/device/sensor topology management, alarm rule evaluation, audit logging, CSV reporting, and controlled commanded pump ON/OFF power operations.\n\n#### Authentication\nUse the `/api/v1/auth/login` endpoint to obtain a JWT Bearer token.\nProvide the token in the header:\n`Authorization: Bearer <your_jwt_token>`\n\n#### Postman Import Instructions\nYou can import this entire API into Postman directly by choosing **Import -> Link** and pasting:\n`http://localhost:3000/docs/json` or downloading the OpenAPI specification from `/docs/json`.\n        ",
    "version": "1.1.0",
    "contact": {
      "name": "PT Ascon Multi Pratama SCADA & Automation Team",
      "url": "https://ascon.co.id"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/alarms/": {
      "get": {
        "summary": "List Alarms",
        "tags": [
          "Alarms"
        ],
        "description": "Returns alarms filtered by site, asset, severity, and status.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            },
            "in": "query",
            "name": "page_size",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sort_by",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "in": "query",
            "name": "sort_order",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "site_id",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "asset_id",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "ACKNOWLEDGED",
                "RESOLVED"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "INFO",
                "LOW",
                "MEDIUM",
                "HIGH",
                "CRITICAL"
              ]
            },
            "in": "query",
            "name": "severity",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/alarms/{id}": {
      "get": {
        "summary": "Get Alarm Detail",
        "tags": [
          "Alarms"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/alarms/{id}/acknowledge": {
      "post": {
        "summary": "Acknowledge Alarm",
        "tags": [
          "Alarms"
        ],
        "description": "Operator acknowledges an active alarm with an optional investigation note.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/alarms/{id}/resolve": {
      "post": {
        "summary": "Manually Resolve Alarm",
        "tags": [
          "Alarms"
        ],
        "description": "Engineer/Admin manually marks an alarm as resolved with explanation.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
...
}